Make list molding work
I have wanted to throw my computer on the floor at least 3 times in the last hour.

Frick this function.

+++ shared/src/main/scala/NumberHelpers.scala
@@ -4,10 +4,10 @@ object NumberHelpers {
 
   def multiplicity(a: VNum, b: VNum): VNum = {
     var result = 0
-    var current = a
-    while (current % b == 0) {
+    var current = a.toInt
+    while (current % b.toInt == 0) {
       result += 1
-      current /= b
+      current /= b.toInt
     }
     result
   }
